:truck: functions folder renamed to lambda
