--- src/functions/hello.js
@@ -1,10 +1,0 @@
-const hello = () => {
-  return 'Hello, world!'
-}
-
-exports.handler = (event, context, callback) => {
-  callback(null, {
-    statusCode: 200,
-    body: hello()
-  })
-}
