+++ src/lambda/config/index.js
