+++ src/lambda/getChartInfo.js
